Move EthApiTypes to new module
